Fix S2139: Exceptions should be either logged or rethrown but not both.
When an exception is logged and rethrown, the upstream code may not be aware that the exception has already been logged. As a result, the same exception gets logged multiple times, making it difficult to identify the root cause of the issue. This can be particularly problematic in multi-threaded applications where messages from other threads can be interwoven with the repeated log entries.

https://sonarcloud.io/project/issues?open=AY8TwSfaw0OXT0TzMOG3&id=feliperomero3_MoneySmart